auto unlock bugfix

+++ ldap_peoples/auth.py
@@ -25,7 +25,7 @@ class LdapAcademiaAuthBackend(ModelBackend):
         ldap_conn = connections['ldap']
         user = None
         lu = LdapAcademiaUser.objects.filter(uid=username).first()
-        if not lu or not lu.is_active():
+        if not lu:
             return None
 
         # check if username exists and if it is active
@@ -38,25 +38,32 @@ class LdapAcademiaAuthBackend(ModelBackend):
             print(e)
             return None
 
+        # if account beign unlocked this will be always false
+        if not lu.is_active():
+            return None
+        
         scoped_username = '@'.join((lu.uid, settings.LDAP_BASE_DOMAIN))
         try:
             user = User.objects.get(username=scoped_username)
-            # update attrs: TODO
-            # if user.email != lu.mail[0]:
-                # DO THINGS
+            # update attrs:
+            if user.email != lu.mail[0]:
+                user.email = lu.mail[0]
+                user.save()                
         except Exception as e:
             user = User.objects.create(dn=lu.dn,
                                        username=scoped_username,
                                        email=lu.mail[0],
                                        first_name=lu.cn,
                                        last_name=lu.sn)
+        
         # disconnect already created session, only a session per user is allowed
         # get all the active sessions
-        for session in Session.objects.all():
-            try:
-                if int(session.get_decoded()['_auth_user_id']) == user.pk:
-                    session.delete()
-            except (KeyError, TypeError, ValueError):
-                pass
+        if not settings.MULTIPLE_USER_AUTH_SESSIONS:
+            for session in Session.objects.all():
+                try:
+                    if int(session.get_decoded()['_auth_user_id']) == user.pk:
+                        session.delete()
+                except (KeyError, TypeError, ValueError):
+                    pass
         
         return user
